Redudent code plus Eye height
removed some redudent code and fixed alpaca eye height for different ages

+++ src/main/java/net/dries007/tfc/objects/blocks/wood/BlockLogPile.java
@@ -102,7 +102,7 @@ public class BlockLogPile extends Block implements ILightableBlock
                 0f, 0.1f + 0.1f * rand.nextFloat(), 0f);
             if (rand.nextDouble() < 0.4D)
             {
-                worldIn.playSound((double) pos.getX() + 0.5D, (double) pos.getY(), (double) pos.getZ() + 0.5D, SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 0.5F, 0.6F, false);
+                worldIn.playSound((double) pos.getX() + 0.5D, pos.getY(), (double) pos.getZ() + 0.5D, SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 0.5F, 0.6F, false);
             }
         }
     }
